+++ cobj/codegen.c
@@ -1408,6 +1408,7 @@ static void joutput_param(cb_tree x, int id) {
   struct cb_alphabet_name *abp;
   struct cb_alphabet_name *rbp;
   cb_tree l;
+  struct literal_list *ll;
   int n;
   int extrefs;
   int sav_stack_id;
@@ -1490,8 +1491,8 @@ static void joutput_param(cb_tree x, int id) {
     joutput("%s%s", CB_PREFIX_FILE, CB_FILE(x)->cname);
     break;
   case CB_TAG_LITERAL:
-    struct literal_list *l = lookup_literal(x);
-    joutput_const_identifier(l);
+    ll = lookup_literal(x);
+    joutput_const_identifier(ll);
     break;
   case CB_TAG_FIELD:
     /* TODO: remove me */
@@ -4012,13 +4013,20 @@ static void joutput_file_initialization(struct cb_file *f) {
           joutput_line ("  byte_%s%s[i] = '0;'");
           joutput_line ("}");
   }*/
+  if (f->external) {
+    joutput_line("%s%s = CobolFile.getExternalFile(\"%s\");", CB_PREFIX_FILE,
+                 f->cname, f->cname);
+    joutput_line("if(%s%s == null) {", CB_PREFIX_FILE, f->cname);
+    joutput_indent_level += 2;
+  }
 
   joutput_line("%s%s = CobolFileFactory.makeCobolFileInstance(", CB_PREFIX_FILE,
                f->cname);
   joutput_line("/* select_name = */ \"%s\",", f->name);
   if (f->external && !f->file_status) {
-    joutput_line("/* file_status = */ cob_external_addr (\"%s%s_status\", 4),",
-                 CB_PREFIX_FILE, f->cname);
+    joutput_line("/* file_status = */ CobolFile.getExternalFileStatus "
+                 "(\"%s\"),",
+                 f->cname);
   } else {
     joutput_line("/* file_status = */ %s%s_status,", CB_PREFIX_FILE, f->cname);
   }
@@ -4076,26 +4084,15 @@ static void joutput_file_initialization(struct cb_file *f) {
   joutput_line(");");
 
   if (f->external) {
-    // joutput_line ("%s%s = CobolExternal.getFileAddress (\"%s\");",
-    // CB_PREFIX_FILE, f->cname, f->cname);
-    joutput_line("if (CobolExternal.initialExternal)");
-    joutput_indent("{");
-    if (f->linage) {
-      joutput_line("%s%s.setLinorkeyptr(new Linage());", CB_PREFIX_FILE,
-                   f->cname);
-    }
-  } else {
-    // joutput_line ("if (%s%s == null)", CB_PREFIX_FILE, f->cname);
-    // joutput_indent ("{");
-    // joutput_line ("%s%s = new CobolFile();", CB_PREFIX_FILE, f->cname);
-    if (f->linage) {
-      joutput_line("%s%s.setLinorkeyptr(new Linage());", CB_PREFIX_FILE,
-                   f->cname);
-    }
-    // joutput_indent ("}");
+    joutput_line("CobolFile.putExternalFile(\"%s\", %s%s);", f->cname,
+                 CB_PREFIX_FILE, f->cname);
+    joutput_indent_level -= 2;
+    joutput_line("}");
   }
 
   if (f->linage) {
+    joutput_line("%s%s.setLinorkeyptr(new Linage());", CB_PREFIX_FILE,
+                 f->cname);
     joutput_line("lingptr = (Linage)(%s%s.getLinorkeyptr());", CB_PREFIX_FILE,
                  f->cname);
     joutput_prefix();
@@ -4134,9 +4131,6 @@ static void joutput_file_initialization(struct cb_file *f) {
     joutput_line("lingptr.setLinFoot(0);");
     joutput_line("lingptr.setLinTop(0);");
     joutput_line("lingptr.setLinBot(0);");
-  }
-  if (f->external) {
-    joutput_indent("}");
   }
 }