+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+## [1.0.15] - 2023-09-29
+### Added
+- Add new intrinsic functions
+  - EXCEPTION-FILE (#218)
+  - EXCEPTION-LOCATION (218)
+  - EXCEPTION-STATEMENT (#221)
+  - EXCEPTION-STATUS (#221)
+  - FRACTION-PART (#224)
+- Add `-Wredefinition` option (#223)
+- Implement `FD` with `EXTERNAL` clause (#222)
+- Add a guideline for contributing (#226)
+### Fixed
+- Fix a build error on some platforms (#225)
 
 ## [1.0.14] - 2023-08-31
 ### Added